Removed CDW stored procedure support for Observation controller
