+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/observation/Dstu2ObservationTransformer.java
@@ -30,7 +30,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Builder
-final class DatamartObservationTransformer {
+final class Dstu2ObservationTransformer {
   @NonNull final DatamartObservation datamart;
 
   static CodeableConcept category(DatamartObservation.Category category) {
