+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/observation/Dstu2ObservationControllerTest.java
@@ -1,6 +1,6 @@
 package gov.va.api.health.dataquery.service.controller.observation;
 
-import static gov.va.api.health.dataquery.service.controller.observation.DatamartObservationSamples.Fhir.link;
+import static gov.va.api.health.dataquery.service.controller.observation.ObservationSamples.Dstu2.link;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -13,11 +13,12 @@ import gov.va.api.health.argonaut.api.resources.Observation.Bundle;
 import gov.va.api.health.autoconfig.configuration.JacksonConfig;
 import gov.va.api.health.dataquery.service.controller.ConfigurableBaseUrlPageLinks;
 import gov.va.api.health.dataquery.service.controller.Dstu2Bundler;
+import gov.va.api.health.dataquery.service.controller.Dstu2Validator;
 import gov.va.api.health.dataquery.service.controller.ResourceExceptions;
 import gov.va.api.health.dataquery.service.controller.WitnessProtection;
 import gov.va.api.health.dataquery.service.controller.observation.DatamartObservation.Category;
-import gov.va.api.health.dataquery.service.controller.observation.DatamartObservationSamples.Datamart;
-import gov.va.api.health.dataquery.service.controller.observation.DatamartObservationSamples.Fhir;
+import gov.va.api.health.dataquery.service.controller.observation.ObservationSamples.Datamart;
+import gov.va.api.health.dataquery.service.controller.observation.ObservationSamples.Dstu2;
 import gov.va.api.health.dstu2.api.bundle.BundleLink.LinkRelation;
 import gov.va.api.health.dstu2.api.datatypes.CodeableConcept;
 import gov.va.api.health.dstu2.api.datatypes.Coding;
@@ -26,6 +27,7 @@ import gov.va.api.health.ids.api.Registration;
 import gov.va.api.health.ids.api.ResourceIdentity;
 import java.time.Instant;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -43,7 +45,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @DataJpaTest
 @RunWith(SpringRunner.class)
-public class DatamartObservationControllerTest {
+public class Dstu2ObservationControllerTest {
 
   HttpServletResponse response;
 
@@ -76,11 +78,8 @@ public class DatamartObservationControllerTest {
     response = mock(HttpServletResponse.class);
   }
 
-  ObservationController controller() {
-    return new ObservationController(
-        true,
-        null,
-        null,
+  Dstu2ObservationController controller() {
+    return new Dstu2ObservationController(
         new Dstu2Bundler(new ConfigurableBaseUrlPageLinks("http://fonzy.com", "cool")),
         repository,
         WitnessProtection.builder().identityService(ids).build());
@@ -102,7 +101,7 @@ public class DatamartObservationControllerTest {
   }
 
   private Multimap<String, Observation> populateData() {
-    var fhir = Fhir.create();
+    var fhir = Dstu2.create();
     var datamart = Datamart.create();
     var observationsByPatient = LinkedHashMultimap.<String, Observation>create();
     var registrations = new ArrayList<Registration>(10);
@@ -158,8 +157,8 @@ public class DatamartObservationControllerTest {
     DatamartObservation dm = Datamart.create().observation();
     repository.save(asEntity(dm));
     mockObservationIdentity("x", dm.cdwId());
-    Observation actual = controller().read("true", "x");
-    assertThat(actual).isEqualTo(DatamartObservationSamples.Fhir.create().observation("x"));
+    Observation actual = controller().read("x");
+    assertThat(actual).isEqualTo(Dstu2.create().observation("x"));
   }
 
   @Test
@@ -187,7 +186,7 @@ public class DatamartObservationControllerTest {
   @Test(expected = ResourceExceptions.NotFound.class)
   public void readThrowsNotFoundWhenDataIsMissing() {
     mockObservationIdentity("x", "x");
-    controller().read("true", "x");
+    controller().read("x");
   }
 
   @Test(expected = ResourceExceptions.NotFound.class)
@@ -200,14 +199,15 @@ public class DatamartObservationControllerTest {
     DatamartObservation dm = Datamart.create().observation();
     repository.save(asEntity(dm));
     mockObservationIdentity("x", dm.cdwId());
-    Bundle actual = controller().searchById("true", "x", 1, 1);
-    Observation observation = Fhir.create().observation("x");
+    Bundle actual = controller().searchById("x", 1, 1);
+    Observation observation = Dstu2.create().observation("x");
     assertThat(json(actual))
         .isEqualTo(
             json(
-                Fhir.asBundle(
+                Dstu2.asBundle(
                     "http://fonzy.com/cool",
                     List.of(observation),
+                    1,
                     link(
                         LinkRelation.first, "http://fonzy.com/cool/Observation?identifier=x", 1, 1),
                     link(LinkRelation.self, "http://fonzy.com/cool/Observation?identifier=x", 1, 1),
@@ -217,23 +217,47 @@ public class DatamartObservationControllerTest {
                         1,
                         1))));
     /* searchById and searchByIdentifier are the same */
-    assertThat(controller().searchByIdentifier("true", "x", 1, 1)).isEqualTo(actual);
+    assertThat(controller().searchByIdentifier("x", 1, 1)).isEqualTo(actual);
   }
 
   @Test(expected = ResourceExceptions.NotFound.class)
   public void searchByIdWhenIdIsUnknown() {
-    controller().searchById("true", "x", 1, 1);
+    controller().searchById("x", 1, 1);
+  }
+
+  @Test
+  public void searchByIdWith0Count() {
+    DatamartObservation dm = Datamart.create().observation();
+    repository.save(asEntity(dm));
+    mockObservationIdentity("x", dm.cdwId());
+    Bundle actual = controller().searchById("x", 1, 0);
+    Observation observation = Dstu2.create().observation("x");
+    assertThat(json(actual))
+        .isEqualTo(
+            json(
+                Dstu2.asBundle(
+                    "http://fonzy.com/cool",
+                    Collections.emptyList(),
+                    1,
+                    link(
+                        LinkRelation.self,
+                        "http://fonzy.com/cool/Observation?identifier=x",
+                        1,
+                        0))));
+    /* searchById and searchByIdentifier are the same */
+    assertThat(controller().searchByIdentifier("x", 1, 0)).isEqualTo(actual);
   }
 
   @Test
   public void searchByPatient() {
     Multimap<String, Observation> observationsByPatient = populateData();
-    assertThat(json(controller().searchByPatient("true", "p0", 1, 10)))
+    assertThat(json(controller().searchByPatient("p0", 1, 10)))
         .isEqualTo(
             json(
-                Fhir.asBundle(
+                Dstu2.asBundle(
                     "http://fonzy.com/cool",
                     observationsByPatient.get("p0"),
+                    observationsByPatient.get("p0").size(),
                     link(LinkRelation.first, "http://fonzy.com/cool/Observation?patient=p0", 1, 10),
                     link(LinkRelation.self, "http://fonzy.com/cool/Observation?patient=p0", 1, 10),
                     link(
@@ -246,18 +270,19 @@ public class DatamartObservationControllerTest {
   @Test
   public void searchByPatientAndCategory() {
     Multimap<String, Observation> observationsByPatient = populateData();
-    assertThat(
-            json(controller().searchByPatientAndCategory("true", "p0", "laboratory", null, 1, 10)))
+    List<Observation> observations =
+        observationsByPatient
+            .get("p0")
+            .stream()
+            .filter(c -> "laboratory".equalsIgnoreCase(c.category().coding().get(0).code()))
+            .collect(Collectors.toList());
+    assertThat(json(controller().searchByPatientAndCategory("p0", "laboratory", null, 1, 10)))
         .isEqualTo(
             json(
-                Fhir.asBundle(
+                Dstu2.asBundle(
                     "http://fonzy.com/cool",
-                    observationsByPatient
-                        .get("p0")
-                        .stream()
-                        .filter(
-                            c -> "laboratory".equalsIgnoreCase(c.category().coding().get(0).code()))
-                        .collect(Collectors.toList()),
+                    observations,
+                    observations.size(),
                     link(
                         LinkRelation.first,
                         "http://fonzy.com/cool/Observation?category=laboratory&patient=p0",
@@ -273,19 +298,19 @@ public class DatamartObservationControllerTest {
                         "http://fonzy.com/cool/Observation?category=laboratory&patient=p0",
                         1,
                         10))));
-    assertThat(
-            json(controller().searchByPatientAndCategory("true", "p1", "vital-signs", null, 1, 10)))
+    List<Observation> patient1Observations =
+        observationsByPatient
+            .get("p1")
+            .stream()
+            .filter(c -> "vital-signs".equalsIgnoreCase(c.category().coding().get(0).code()))
+            .collect(Collectors.toList());
+    assertThat(json(controller().searchByPatientAndCategory("p1", "vital-signs", null, 1, 10)))
         .isEqualTo(
             json(
-                Fhir.asBundle(
+                Dstu2.asBundle(
                     "http://fonzy.com/cool",
-                    observationsByPatient
-                        .get("p1")
-                        .stream()
-                        .filter(
-                            c ->
-                                "vital-signs".equalsIgnoreCase(c.category().coding().get(0).code()))
-                        .collect(Collectors.toList()),
+                    patient1Observations,
+                    patient1Observations.size(),
                     link(
                         LinkRelation.first,
                         "http://fonzy.com/cool/Observation?category=vital-signs&patient=p1",
@@ -343,20 +368,22 @@ public class DatamartObservationControllerTest {
     testDates.putAll("gt2005-01-14", List.of());
     testDates.putAll("sa2005-01-14", List.of());
     for (var date : testDates.keySet()) {
+      List<Observation> observations =
+          observationsByPatient
+              .get("p0")
+              .stream()
+              .filter(o -> testDates.get(date).contains(o.effectiveDateTime()))
+              .collect(Collectors.toList());
       assertThat(
               json(
                   controller()
-                      .searchByPatientAndCategory(
-                          "true", "p0", "laboratory", new String[] {date}, 1, 10)))
+                      .searchByPatientAndCategory("p0", "laboratory", new String[] {date}, 1, 10)))
           .isEqualTo(
               json(
-                  Fhir.asBundle(
+                  Dstu2.asBundle(
                       "http://fonzy.com/cool",
-                      observationsByPatient
-                          .get("p0")
-                          .stream()
-                          .filter(o -> testDates.get(date).contains(o.effectiveDateTime()))
-                          .collect(Collectors.toList()),
+                      observations,
+                      observations.size(),
                       link(
                           LinkRelation.first,
                           "http://fonzy.com/cool/Observation?category=laboratory&date="
@@ -412,11 +439,16 @@ public class DatamartObservationControllerTest {
         List.of("2005-01-10T07:57:00Z", "2005-01-12T07:57:00Z", "2005-01-14T07:57:00Z"));
     testDates.putAll(Pair.of("gt2005-01-13", "lt2005-01-15"), List.of("2005-01-14T07:57:00Z"));
     for (var date : testDates.keySet()) {
+      List<Observation> observations =
+          observationsByPatient
+              .get("p0")
+              .stream()
+              .filter(o -> testDates.get(date).contains(o.effectiveDateTime()))
+              .collect(Collectors.toList());
       assertThat(
               json(
                   controller()
                       .searchByPatientAndCategory(
-                          "true",
                           "p0",
                           "laboratory",
                           new String[] {date.getLeft(), date.getRight()},
@@ -424,13 +456,10 @@ public class DatamartObservationControllerTest {
                           10)))
           .isEqualTo(
               json(
-                  Fhir.asBundle(
+                  Dstu2.asBundle(
                       "http://fonzy.com/cool",
-                      observationsByPatient
-                          .get("p0")
-                          .stream()
-                          .filter(o -> testDates.get(date).contains(o.effectiveDateTime()))
-                          .collect(Collectors.toList()),
+                      observations,
+                      observations.size(),
                       link(
                           LinkRelation.first,
                           "http://fonzy.com/cool/Observation?category=laboratory&date="
@@ -464,16 +493,19 @@ public class DatamartObservationControllerTest {
   @Test
   public void searchByPatientAndCode() {
     Multimap<String, Observation> observationsByPatient = populateData();
-    assertThat(json(controller().searchByPatientAndCode("true", "p0", "1989-3", 1, 10)))
+    List<Observation> patient0Observations =
+        observationsByPatient
+            .get("p0")
+            .stream()
+            .filter(c -> "1989-3".equalsIgnoreCase(c.code().coding().get(0).code()))
+            .collect(Collectors.toList());
+    assertThat(json(controller().searchByPatientAndCode("p0", "1989-3", 1, 10)))
         .isEqualTo(
             json(
-                Fhir.asBundle(
+                Dstu2.asBundle(
                     "http://fonzy.com/cool",
-                    observationsByPatient
-                        .get("p0")
-                        .stream()
-                        .filter(c -> "1989-3".equalsIgnoreCase(c.code().coding().get(0).code()))
-                        .collect(Collectors.toList()),
+                    patient0Observations,
+                    patient0Observations.size(),
                     link(
                         LinkRelation.first,
                         "http://fonzy.com/cool/Observation?code=1989-3&patient=p0",
@@ -489,16 +521,19 @@ public class DatamartObservationControllerTest {
                         "http://fonzy.com/cool/Observation?code=1989-3&patient=p0",
                         1,
                         10))));
-    assertThat(json(controller().searchByPatientAndCode("true", "p1", "8480-6", 1, 10)))
+    List<Observation> patient1Observations =
+        observationsByPatient
+            .get("p1")
+            .stream()
+            .filter(c -> "8480-6".equalsIgnoreCase(c.code().coding().get(0).code()))
+            .collect(Collectors.toList());
+    assertThat(json(controller().searchByPatientAndCode("p1", "8480-6", 1, 10)))
         .isEqualTo(
             json(
-                Fhir.asBundle(
+                Dstu2.asBundle(
                     "http://fonzy.com/cool",
-                    observationsByPatient
-                        .get("p1")
-                        .stream()
-                        .filter(c -> "8480-6".equalsIgnoreCase(c.code().coding().get(0).code()))
-                        .collect(Collectors.toList()),
+                    patient1Observations,
+                    patient1Observations.size(),
                     link(
                         LinkRelation.first,
                         "http://fonzy.com/cool/Observation?code=8480-6&patient=p1",
@@ -552,11 +587,16 @@ public class DatamartObservationControllerTest {
         Pair.of("gt2005-01-13", "lt2005-01-17"),
         List.of("2005-01-14T07:57:00Z", "2005-01-16T07:57:00Z"));
     for (var date : testDates.keySet()) {
+      List<Observation> observations =
+          observationsByPatient
+              .get("p0")
+              .stream()
+              .filter(o -> testDates.get(date).contains(o.effectiveDateTime()))
+              .collect(Collectors.toList());
       assertThat(
               json(
                   controller()
                       .searchByPatientAndCategory(
-                          "true",
                           "p0",
                           "laboratory,vital-signs",
                           new String[] {date.getLeft(), date.getRight()},
@@ -564,13 +604,10 @@ public class DatamartObservationControllerTest {
                           10)))
           .isEqualTo(
               json(
-                  Fhir.asBundle(
+                  Dstu2.asBundle(
                       "http://fonzy.com/cool",
-                      observationsByPatient
-                          .get("p0")
-                          .stream()
-                          .filter(o -> testDates.get(date).contains(o.effectiveDateTime()))
-                          .collect(Collectors.toList()),
+                      observations,
+                      observations.size(),
                       link(
                           LinkRelation.first,
                           "http://fonzy.com/cool/Observation?category=laboratory,vital-signs&date="
@@ -599,5 +636,33 @@ public class DatamartObservationControllerTest {
                           1,
                           10))));
     }
+  }
+
+  @Test
+  public void validate() {
+    Observation observation = ObservationSamples.Dstu2.create().observation("1", "12345V67890");
+    assertThat(
+            controller()
+                .validate(
+                    ObservationSamples.Dstu2.asBundle(
+                        "http://fonzy.com/cool",
+                        List.of(observation),
+                        1,
+                        Dstu2.link(
+                            LinkRelation.first,
+                            "http://fonzy.com/cool/Observation?identifier=1",
+                            1,
+                            1),
+                        Dstu2.link(
+                            LinkRelation.self,
+                            "http://fonzy.com/cool/Observation?identifier=1",
+                            1,
+                            1),
+                        Dstu2.link(
+                            LinkRelation.last,
+                            "http://fonzy.com/cool/Observation?identifier=1",
+                            1,
+                            1))))
+        .isEqualTo(Dstu2Validator.ok());
   }
 }
