--- data-query/src/test/resources/cdw/old-observation-1.04.json
@@ -1,51 +1,0 @@
-{
-  "resourceType" : "Observation",
-  "id" : "8ae37ef2-0047-50b5-889a-d849f3c5d959",
-  "status" : "final",
-  "category" : {
-    "coding" : [
-      {
-        "system" : "http://hl7.org/fhir/observation-category",
-        "code" : "laboratory",
-        "display" : "Laboratory"
-      }
-    ]
-  },
-  "code" : {
-    "coding" : [
-      {
-        "system" : "http://loinc.org",
-        "code" : "786-4",
-        "display" : "MCHC"
-      }
-    ]
-  },
-  "subject" : {
-    "reference" : "https://www.freedomstream.io/Argonaut/api/Patient/185601V825290",
-    "display" : "VETERAN,JOHN Q"
-  },
-  "effectiveDateTime" : "2015-04-15T15:30:02Z",
-  "issued" : "2015-04-15T16:20:02Z",
-  "valueQuantity" : {
-    "value" : 33.6,
-    "unit" : "g/dL",
-    "system" : "http://unitsofmeasure.org",
-    "code" : "g/dL"
-  },
-  "referenceRange" : [
-    {
-      "high" : {
-        "value" : 37.1,
-        "unit" : "g/dL",
-        "system" : "http://unitsofmeasure.org",
-        "code" : "g/dL"
-      },
-      "low" : {
-        "value" : 31.8,
-        "unit" : "g/dL",
-        "system" : "http://unitsofmeasure.org",
-        "code" : "g/dL"
-      }
-    }
-  ]
-}
